docs: tweaks to docs refactor: repackage
